Magical/Item continuation

+++ Scripts/Model/Item/Item.cs
@@ -6,15 +6,15 @@ public abstract class Item : ScriptableObject{
 
     [SerializeField] private string _name;
     [TextArea] private string _description;
-    [SerializeField] private int _weight;
+    [SerializeField] private double _weight;
     [SerializeField] private Rarity _rarity;
 
     public string Name { get=> _name;}
     public string Description {get=> _description; set=> _description=value;}
-    public int Weight {get=> _weight;}
+    public double Weight {get=> _weight;}
     public Rarity Rarity{get=> _rarity;}
 
-    public Item(string name,string desc,int weight,Rarity rar){
+    public Item(string name,string desc,double weight,Rarity rar){
         _name = name;
         _description = desc;
         _weight = weight;
